login and signup linked with database

+++ creating-joy/src/main/java/com/technocrats/creatingjoy/service/UserServiceImpl.java
@@ -1,30 +1,27 @@
 package com.technocrats.creatingjoy.service;
 
-import com.technocrats.creatingjoy.dao.RoleRepository;
 import com.technocrats.creatingjoy.dao.UserRepository;
-import com.technocrats.creatingjoy.entity.Role;
+import com.technocrats.creatingjoy.dto.UserDTO;
 import com.technocrats.creatingjoy.entity.User;
+import com.technocrats.creatingjoy.objectmapper.UserMapper;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Collection;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Service
+@Slf4j
 public class UserServiceImpl implements UserService {
 
     private UserRepository userRepository;
 
     @Autowired
-    private RoleRepository roleRepository;
+    private UserMapper userMapper;
+
 
     @Autowired
     public UserServiceImpl(UserRepository userRepository) {
@@ -35,67 +32,53 @@ public class UserServiceImpl implements UserService {
     @Autowired
     private BCryptPasswordEncoder passwordEncoder;
 
-    @Override
-    @Transactional
-    public List<User> findAll() {
 
-        return userRepository.findAll();
+    @Override
+    public List<UserDTO> findAll() {
+
+        List<User> allUsers = new ArrayList<>();
+        allUsers = userRepository.findAll();
+        List<UserDTO> allUsersDto = new ArrayList<>();
+        for(User user: allUsers)
+            allUsersDto.add(userMapper.convertToDto(user));
+
+        return allUsersDto;
     }
 
     @Override
-    @Transactional
-    public User findById(int id) {
+    public UserDTO findById(int id) {
 
-        Optional<User> result=userRepository.findById(id);
-
-        if(result.isPresent()){
-            return result.get();
+        Optional<User> user = userRepository.findById(id);
+        if(user.isPresent()){
+            UserDTO userDto = userMapper.convertToDto(user.get());
+            return userDto;
         }
-        else{
-            throw new RuntimeException("Did not found User with id "+id);
-        }
-
-
+        else
+            throw new RuntimeException("User not found :" + id);
     }
 
     @Override
-    @Transactional
-    public User findByUserName(String userName) {
-        // check the database if the user already exists
-        return userRepository.findByUserName(userName);
-    }
-
-    @Override
-    @Transactional
-    public void save(User user) {
-
-
-
-        user.setPassword(passwordEncoder.encode(user.getPassword()));
+    public void save(UserDTO userDto) {
+        User user = userMapper.convertToEntity(userDto);
         userRepository.save(user);
-
     }
 
     @Override
-    @Transactional
     public void deleteById(int id) {
+        if(id>0)
+            userRepository.deleteById(id);
+    }
 
-        userRepository.deleteById(id);
+
+    public UserDTO findByUserName(String userName) {
+        // check the database if the user already exists
+        User user =  userRepository.findByUserName(userName);
+        UserDTO userDto = null;
+        if(user!=null)
+            userDto = userMapper.convertToDto(user);
+        return userDto;
 
     }
 
-    @Override
-    @Transactional
-    public UserDetails loadUserByUsername(String userName) throws UsernameNotFoundException {
-        User user = userRepository.findByUserName(userName);
-        if (user == null) {
-            throw new UsernameNotFoundException("Invalid username or password.");
-        }
-        return new org.springframework.security.core.userdetails.User(user.getUserName(), user.getPassword(),
-                mapRolesToAuthorities(user.getRoles()));
-    }
 
-    private Collection<? extends GrantedAuthority> mapRolesToAuthorities(Collection<Role> roles) {
-        return roles.stream().map(role -> new SimpleGrantedAuthority(role.getName())).collect(Collectors.toList());
-    }
 }
